requirements.txt
discord.py
selenium
webdriver-manager
Pillow
python-dotenv